requirements.txt
fastapi
uvicorn
jinja2
google-cloud-bigquery
